requirements.txt
python-dotenv
matplotlib
Pillow
ffmpeg-python